Add a Microdata parsing filter using Any23
This is a ParseFilter which uses Any23 to perform the microdata parsing
It adds to the metadata all the encountered microdata using a key-dotted
path. There are 2 modes toggled via a boolean
"useUniquePrefixForNestedValues". When set to true, nested values get a
unique numerical prefix. The defaut is to merge all of the values within
a single entry instead. Not sure yet which is better.

I added a few tests to parse pages I found on the web using microdata.