BE-472 #Done Put Errors in the status stream in JsoupParser Bolt
Currently the error handling of the JSoupParserBolt is incorrect, It is currently doing a collector.fail,
unfortunately this just resends the same tuple over and over again when we do this.

A better solution would be to send the crawlStatus through the status stream to the CrawlItemUpdater.
The crawlItemUpdater will then save the status to dynamoDB with a status of error.