Fix NPE in JSoupParserBolt
When emitting outlinks, we hadn't accounted for the fact that the URL
is normalized by the filters. So when retrieving the anchors, the key
was no longer the original one. Ensuring that we retrieve the anchors
at the same time we retrive the outlink URL, guarantees that we won't
get the NPE